requirements.txt
aiohttp
astroid
async-timeout
attrs
cffi
chardet
colorama
cryptography
discord.py
ffmpeg
idna
isort
lazy-object-proxy
mccabe
multidict
pycparser
pylint
pymongo
PyNaCl
pyspnego
six
smbprotocol
toml
typing_extensions
wrapt
yarl==1.8.2
yt-dlp
requests
python-dotenv
python-dateutil